refactor: moves searching algos into own folder
