+++ CONTENTS.md
@@ -4,3 +4,15 @@
 
 -   [Singly Linked List](./src/data-structures/linked-lists/SinglyLinkedList.ts)
 -   [Doubly Linked List](./src/data-structures/linked-lists/DoublyLinkedList.ts)
+
+## Algorithms
+
+### Sorting
+
+-   [Bubble Sort](./src/algorithms/sorting/bubbleSort.ts)
+-   [Insertion Sort](./src/algorithms/sorting/insertionSort.ts)
+
+### Searching
+
+-   [Binary Search](./src/algorithms/searching/binarySearch.ts)
+-   [Linear Search](./src/algorithms/searching/linearSearch.ts)
