+++ src/algorithms/searching/binarySearch.ts
@@ -1,12 +1,3 @@
-type valueType = string | number;
-
-export function linearSearch(arr: valueType[], value: valueType): boolean {
-    for (let index = 0; index < arr.length; index++) {
-        if (arr[index] === value) return true;
-    }
-    return false;
-}
-
 export function binarySearch(arr: number[], value: number): boolean {
     let sorted = arr.sort((a, b) => a - b);
